Consolidate two extraction-related modules into one
