+++ src/beneficiations/beneficiation.py
@@ -14,7 +14,7 @@ import logging
 from typing import Any, Dict
 
 from src.beneficiations.abstract_custom_beneficiation import AbstractCustomBeneficiation
-from src.extraction_output_manager.ingestibles import IngestibleDataclasses
+from src.extraction.ingestibles import IngestibleDataclasses
 from src.utils.types.singleton import Singleton
 
 # ---Constants
