+++ src/extraction/output_manager.py
