+++ src/miners/abstract_custom_miner.py
@@ -13,7 +13,7 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Optional
 
-from src.extraction_output_manager import output_manager as om
+from src.extraction import output_manager as om
 
 # ---Constants
