+++ src/profiles/abi_music/miner_helpers/dataclass_identifier.py
@@ -16,7 +16,7 @@ from typing import Any, Dict, List, Optional
 
 import yaml
 
-from src.extraction_output_manager import output_manager as om
+from src.extraction import output_manager as om
 
 # ---Constants
 logger = logging.getLogger(__name__)
