+++ src/profiles/abi_music/miner_helpers/datafile_miner.py
@@ -19,7 +19,7 @@ from typing import Any, Dict, Optional
 
 import yaml
 
-from src.extraction_output_manager import output_manager as om
+from src.extraction import output_manager as om
 from src.miners.utils import datafile_metadata_helpers as dmh
 from src.profiles import profile_consts as pc
 from src.profiles import profile_helpers as ph
