+++ src/profiles/abi_music/miner_helpers/project_miner.py
@@ -13,7 +13,7 @@ import logging
 from pathlib import Path
 from typing import Any, Dict, Optional
 
-from src.extraction_output_manager import output_manager as om
+from src.extraction import output_manager as om
 from src.profiles import profile_consts as pc
 from src.profiles.abi_music import abi_music_consts as amc
 from src.profiles.abi_music.miner_helpers import metadata_helpers as mh
