+++ src/profiles/abi_music/prospector_helpers/directory_prospector.py
@@ -14,7 +14,7 @@ import os
 from pathlib import Path
 from typing import Any, Optional
 
-from src.extraction_output_manager import output_manager as om
+from src.extraction import output_manager as om
 from src.profiles import profile_consts as pc
 from src.profiles.abi_music import abi_music_consts as amc
