+++ src/prospectors/abstract_custom_prospector.py
@@ -8,7 +8,7 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Optional
 
-from src.extraction_output_manager import output_manager as om
+from src.extraction import output_manager as om
 
 # ---Constants
 logger = logging.getLogger(__name__)
