+++ src/prospectors/prospector.py
@@ -9,7 +9,7 @@ import logging
 from pathlib import Path
 from typing import Optional, Union
 
-from src.extraction_output_manager import output_manager as om
+from src.extraction import output_manager as om
 from src.prospectors.abstract_custom_prospector import AbstractCustomProspector
 from src.prospectors.common_file_checks import CommonDirectoryTreeChecks
 from src.utils.types.singleton import Singleton
