+++ tests/ignored/ignoredtest_abi_beneficiation.py
@@ -7,8 +7,8 @@ import pytest
 from pytest import fixture
 
 from src.beneficiations.beneficiation import Beneficiation
-from src.extraction_output_manager.ingestibles import IngestibleDataclasses
-from src.extraction_output_manager.output_manager import OutputManager
+from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.output_manager import OutputManager
 from src.profiles import profile_consts as pc
 from src.profiles.profile_loader import ProfileLoader
 from tests.fixtures.fixtures_abi_data import (
