+++ tests/test_abi_beneficiations.py
@@ -8,8 +8,8 @@ from pytest import fixture
 
 from src.beneficiations import beneficiation_consts as bc
 from src.beneficiations.beneficiation import Beneficiation
-from src.extraction_output_manager.ingestibles import IngestibleDataclasses
-from src.extraction_output_manager.output_manager import OutputManager
+from src.extraction.ingestibles import IngestibleDataclasses
+from src.extraction.output_manager import OutputManager
 from src.profiles import profile_consts as pc
 from tests.fixtures.fixtures_abi_data import (
     get_bad_beneficiation_format,
